Split telemetry context helpers
